pyproject.toml
[project]
name = "tevm"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pywebview (>=6.1,<7.0)"
]

[tool.poetry]
packages = [{include = "tevm", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
